Refactored generating link logic
