+++ backend/EasyMeets.Core/EasyMeets.Core.BLL/Interfaces/ITeamSharedService.cs
@@ -6,6 +6,4 @@ namespace EasyMeets.Core.BLL.Interfaces;
 public interface ITeamSharedService
 {
     Task CreateDefaultUsersTeamAsync(User user);
-    Task<string> GenerateNewPageLinkAsync(long teamId, string teamName);
-    string GenerateInvivationLink(IUrlHelper Url, long? userId, string userEmail, long teamId);
 }
