+++ backend/EasyMeets.Core/EasyMeets.Core.BLL/Services/TeamSharedService.cs
@@ -1,22 +1,17 @@
 using AutoMapper;
 using EasyMeets.Core.BLL.Interfaces;
-using EasyMeets.Core.Common.DTO.Common;
 using EasyMeets.Core.Common.DTO.Team;
 using EasyMeets.Core.Common.Enums;
 using EasyMeets.Core.DAL.Context;
 using EasyMeets.Core.DAL.Entities;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using Newtonsoft.Json;
-using System.Text;
-using System.Web;
-
 namespace EasyMeets.Core.BLL.Services;
 
 public class TeamSharedService : BaseService, ITeamSharedService
 {
-    public TeamSharedService(EasyMeetsCoreContext context, IMapper mapper) : base(context, mapper)
+    private readonly ILinkService _linkService;
+    public TeamSharedService(EasyMeetsCoreContext context, IMapper mapper, ILinkService linkService) : base(context, mapper)
     {
+        _linkService = linkService;
     }
 
     public async Task CreateDefaultUsersTeamAsync(User user)
@@ -25,7 +20,7 @@ public class TeamSharedService : BaseService, ITeamSharedService
         var teamDto = new TeamDto
         {
             Name = teamName,
-            PageLink = await GenerateNewPageLinkAsync(0, teamName),
+            PageLink = await _linkService.GenerateNewPageLinkAsync(0, teamName),
         };
 
         var team = _mapper.Map<Team>(teamDto);
@@ -41,34 +36,5 @@ public class TeamSharedService : BaseService, ITeamSharedService
 
         _context.TeamMembers.Add(member);
         await _context.SaveChangesAsync();
-    }
-
-    public async Task<string> GenerateNewPageLinkAsync(long teamId, string teamName)
-    {
-        if (!await _context.Teams.AnyAsync(t => t.Id != teamId && t.PageLink == teamName))
-        {
-            return teamName;
-        }
-
-        int index = 1;
-
-        while (await _context.Teams.AnyAsync(t => t.Id != teamId && t.PageLink == $"{teamName}{index}"))
-        {
-            index++;
-        }
-
-        return $"{teamName}{index}";
-    }
-    public string GenerateInvivationLink(IUrlHelper Url, long? userId, string userEmail, long teamId)
-    {
-        var teamDataJson = userId == null ?
-                JsonConvert.SerializeObject(new UserInvitationDataDto { UserEmail = userEmail, TeamId = teamId }) :
-                JsonConvert.SerializeObject(new UserInvitationDataDto { UserId = userId, UserEmail = userEmail, TeamId = teamId });
-
-        string urlEncodedTeamData = HttpUtility.UrlEncode(teamDataJson, Encoding.UTF8);
-
-        var url = Url.Link("AcceptInvitation", new { ecodedTeamData = urlEncodedTeamData });
-
-        return url;
     }
 }
